[Feat] - 유저 계정 인증 API 6개 추가
