+++ app/src/main/java/com/zion830/threedollars/datasource/model/v4/certification/LoginResponse.kt
@@ -1,0 +1,11 @@
+package com.zion830.threedollars.datasource.model.v4.certification
+
+
+import com.google.gson.annotations.SerializedName
+
+data class LoginResponse(
+    @SerializedName("token")
+    val token: String = "",
+    @SerializedName("userId")
+    val userId: Int = 0,
+)
