+++ app/src/main/java/com/zion830/threedollars/datasource/model/v4/certification/request/ConnectAccountRequest.kt
@@ -1,0 +1,6 @@
+package com.zion830.threedollars.datasource.model.v4.certification.request
+
+data class ConnectAccountRequest(
+    val token: String,
+    val socialType: String,
+)
