+++ app/src/main/java/com/zion830/threedollars/datasource/model/v4/certification/request/SignUpRequest.kt
@@ -1,0 +1,13 @@
+package com.zion830.threedollars.datasource.model.v4.certification.request
+
+
+import com.google.gson.annotations.SerializedName
+
+data class SignUpRequest(
+    @SerializedName("name")
+    val name: String = "",
+    @SerializedName("socialType")
+    val socialType: String = "KAKAO",
+    @SerializedName("token")
+    val token: String = ""
+)
